requirements.txt
surfacescan==0.1.0
PyYAML>=6.0
-e ../SurfaceScan/  # Referência local ao projeto SurfaceScan